--- entity-framework/core/saving/explicit-values-generated-properties.md
@@ -1,87 +1,0 @@
----
-title: Setting Explicit Values for Generated Properties - EF Core
-description: Information on setting values explicitly for properties configured as generated with Entity Framework Core
-author: ajcvickers
-ms.date: 10/27/2016
-uid: core/saving/explicit-values-generated-properties
----
-
-# Setting Explicit Values for Generated Properties
-
-A generated property is a property whose value is generated (either by EF or the database) when the entity is added and/or updated. See [Generated Properties](xref:core/modeling/generated-properties) for more information.
-
-There may be situations where you want to set an explicit value for a generated property, rather than having one generated.
-
-> [!TIP]
-> You can view this article's [sample](https://github.com/dotnet/EntityFramework.Docs/tree/master/samples/core/Saving/ExplicitValuesGenerateProperties/) on GitHub.
-
-## The model
-
-The model used in this article contains a single `Employee` entity.
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/Employee.cs#Sample)]
-
-## Saving an explicit value during add
-
-The `Employee.EmploymentStarted` property is configured to have values generated by the database for new entities (using a default value).
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/EmployeeContext.cs#EmploymentStarted)]
-
-The following code inserts two employees into the database.
-
-* For the first, no value is assigned to `Employee.EmploymentStarted` property, so it remains set to the CLR default value for `DateTime`.
-* For the second, we have set an explicit value of `1-Jan-2000`.
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/Sample.cs#EmploymentStarted)]
-
-Output shows that the database generated a value for the first employee and our explicit value was used for the second.
-
-```output
-1: John Doe, 1/26/2017 12:00:00 AM
-2: Jane Doe, 1/1/2000 12:00:00 AM
-```
-
-### Explicit values into SQL Server IDENTITY columns
-
-By convention the `Employee.EmployeeId` property is a store generated `IDENTITY` column.
-
-For most situations, the approach shown above will work for key properties. However, to insert explicit values into a SQL Server `IDENTITY` column, you need to manually enable `IDENTITY_INSERT` before calling `SaveChanges()`.
-
-> [!NOTE]
-> We have a [feature request](https://github.com/aspnet/EntityFramework/issues/703) on our backlog to do this automatically within the SQL Server provider.
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/Sample.cs#EmployeeId)]
-
-Output shows that the supplied ids were saved to the database.
-
-```output
-100: John Doe
-101: Jane Doe
-```
-
-## Setting an explicit value during update
-
-The `Employee.LastPayRaise` property is configured to have values generated by the database during updates.
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/EmployeeContext.cs#LastPayRaise)]
-
-> [!NOTE]
-> By default, EF Core will throw an exception if you try to save an explicit value for a property that is configured to be generated during update. To avoid this, you need to drop down to the lower level metadata API and set the `AfterSaveBehavior` (as shown above).
-
-There is also a trigger in the database to generate values for the `LastPayRaise` column during `UPDATE` operations.
-
-[!code-sql[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/employee_UPDATE.sql)]
-
-The following code increases the salary of two employees in the database.
-
-* For the first, no value is assigned to `Employee.LastPayRaise` property, so it remains set to null.
-* For the second, we have set an explicit value of one week ago (back dating the pay raise).
-
-[!code-csharp[Main](../../../samples/core/Saving/ExplicitValuesGenerateProperties/Sample.cs#LastPayRaise)]
-
-Output shows that the database generated a value for the first employee and our explicit value was used for the second.
-
-```output
-1: John Doe, 1/26/2017 12:00:00 AM
-2: Jane Doe, 1/19/2017 12:00:00 AM
-```
