+++ samples/core/SqlServer/ValueGeneration/Blog.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace SqlServer.ValueGeneration
+{
+    public class Blog
+    {
+        public int BlogId { get; set; }
+        public string Url { get; set; }
+    }
+}
